revert: refactor: use ES6 import syntax
Revert 1cb9b9feb9146d41ca15cb69ce7319a2b9af6c51 because qunit cannot
handle the ES6 import syntax and I don't see a simple way to use webpack
to transform it. Maybe we should consider switching to another unittest
provider in the future, but this is not the focus right now.