feat: resolve internal links before rendering
In order to be able to properly render an internal link, we need more
information from the server. Hence, upon submitting the `LinkForm` for
an internal link, there will be an ajax request retrieving the resolved
link, the automatic heading and whether the page exists.

It might be beneficial to do this during typing, but this optimization
can be done later.