feat: support windows share links as 'other'
Windows share links are too niche to receive their own type for now,
thus they are grouped under 'other'. However the url validation for
external links doesn't allow for addresses of local shares, so I created
the new type which comes without validation on that field.

We may have to extend this type, which currently only covers windows
share links, when it comes to setting the class.