🐛 Don't destroy form if we still need it
As stated in the commit it is not clear when and with what purpose the
destroy method is called. There was a bug that a remaining reference to
the node's attributes object prevented the destroy method to be called.
While this might indicate garbage collection as the trigger, the
LinkView object itself remains in use and it's constructor will not be
called again.