🏗 Show LinkForm before creating the node
Selecting a node after it has been created is tricky. So we show the form
first and create the node if the user-provided data. Afterwards the cursor
is behind the node. (This is analogous to how the insert image menu item in
the example works)

The observed bug in the other approach was, that selecting the node
failed when inserting it at the end of the line. But possibly the actual
bug was, that it worked everywhere else 🤷