🐛 fix syntax allowed in list nodes
This fixes the problem that headlines and quoting was possible in list
nodes. To achieve this we build list nodes closer to how they are in
dokuwiki. They must now contain one list_content node and may contain
one bullet_list or ordered_list node. The list_content node may contain
a list_paragraph (it's just like an ordinary paragraph, except that it
is rendered in `<div>` instead of `<p>` tags to prevent extra styling)
or nodes from the group protected_block or substitution_block. The
groups are created analogous to the PARSER_MODES in DokuWiki, except
that we have to differentiate between block and inline nodes.

This should fix two more items mentioned in #28