🐛 Go back to implementing lists with <p>
The previous approach didn't work.

Using a list_paragraph inside of the li failed, because then Prosemirror
would no longer wrap ordinary paragraphs in list items.
Using an ordinary paragraph inside of a list item wouldn't work because
than splitting list items with the Enter-key breaks because it only
splits list items if one is in the direct descendant of the list_item.

Hence, using paragraphs directly in the list item and fixing styles via
css.