🐛 (Tables) Allow some blocks in table cells
This fixes two things:

1. DokuWiki allows for some blocks in table cells, i.e. protected and
substitution. Since we cannot mix blocks and inline content in
Prosemirror, we have to allow only block-nodes within table cells and
hence must start with a paragraph to support simple text.

When allowing for paragraphs we have to guard our trimming statement
because `ltrim(null)` is apparently an empty string.

2. The splitCell command from prosemirror-tables module implicitly expects to
be called from within a block element in a table cell. Because their
cellWrapping utlility function starts at looking for the table cell at
`pos.depth - 1` which is the table row if the cell only contains inline
content. See also ProseMirror/prosemirror-tables#59 for this.