✨ Implement ~~NOCACHE~~ and ~~NOTOC~~
These macros have no visual representation in the xhtml rendered by
DokuWiki and thus should not be rendered in the prosemirror editor
either.

Attaching this to the root not seems to be cleanest approach. However,
this has some challenges, because by default the attributes on the doc
node cannot be edited because, currently, prosemirror transformations
only apply to the _content_ of the document. However, marijn
indicated[0] that it might be worth adding a step for this to the
official packages.
We are here using a custom step that was provided in another thread[1]
on the official form that does exactly that: setting attributes on the
doc node.

However, these macros make only sense on a top-level page, not in
footnotes. Thus, they are removed from that schema. Something similar
would likely have to be done for a schema for the wrap-plugin or other
similar plugins.

Also, I think checkboxes are the best way to go for representing
toggeling of options. Thus, I added them with a render function and
hide/show them during the update() step. The background is that the
render function is only run during the initialisation and later only
updated via the update() method. See the MenuView class and the called
methods in MenuItem.

Further, the macros themselves are not rendered in the HTML, but the
spaces and linebreaks may still generate instructions for an more or
less empty paragraph. To ensure consistent spacing in the syntax
between the blocks, I've added the rtrim call in the RootNode. Thus
all blocks in the syntax should now always be separated by exactly one
empty line.

[0] https://discuss.prosemirror.net/t/meta-data-nodes/1163/6
[1] https://discuss.prosemirror.net/t/changing-doc-attrs/784/17