✨ Add insertParagraph buttons for html_block
Exemplary implementation of the insertParagraph buttons.

The buttons need access to the current view to execute the command. Due
to javascript scoping rules we have to provide a function to get the view when
we need it instead of a reference to the view itself.

Also, strange things are happening if we create the buttons right in the
element where we need them. Thus creating them in a local variable and
adding them when needed.

ToDo:
* decide where we actually want to use them
* styling