🏗 Use class based approach InsertParagraph buttons
The previous implementation was flawed beyond repair, because there was
only ever a single instance of the buttons. They were in effect only
attached to the very last element of those which were supposed to have
them.

This new implementation lets ProseMirror create the buttons and executes
the event based on their class and get's the view from the global
ProseMirror field.